requirements.txt
openai-whisper>=20231117
pydub>=0.25.1
rich>=13.7.0
numpy>=1.24.0
audioop-lts>=0.2.1
